Move almost all of FormattingUtilities
There was one simple method that was called from the Workspaces layer, but that can just be an extension method
